% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnorm.R
\name{logitnorm}
\alias{logitnorm}
\alias{rlogitnorm}
\alias{dlogitnorm}
\alias{plogitnorm}
\alias{qlogitnorm}
\title{Functions related to logit-normal distribution.}
\usage{
rlogitnorm(n, mu, sigma)

dlogitnorm(x, mu, sigma)

plogitnorm(x, mu, sigma)

qlogitnorm(x, mu, sigma)
}
\arguments{
\item{n}{Number of draws requested (for rlogitnorm)}

\item{mu}{Mean of the logit-transformed variable}

\item{sigma}{SD of the logit-transformed variable}

\item{x}{For density, CDF, and quantile functions}
}
\value{
Depends on the function
}
\description{
Functions related to logit-normal distribution.
}
