% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mcmap}
\alias{mcmap}
\title{The main mapper function}
\usage{
mcmap(target, type = c("beta", "logitnorm", "probitnorm"))
}
\arguments{
\item{target}{A vector of size 2. The first element is mean and the second element is c-statistic.}

\item{type}{One of "beta", "logitnorm", "probitnorm". Loosy matching is enabled (so "b" will be mapped to "beta").}
}
\value{
An object of class mcmapper. The "value" component returns the parameter. Any warning or error from the integration or gradient ascent will also be returned in the "info" component.
}
\description{
Maps a pair of mean and c-statistic value to the parameters of a specified distribution for risk
}
\examples{
mcmap(c(0.1, 0.75), "beta")
}
