% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{mcmap_beta}
\alias{mcmap_beta}
\title{Mapper function for beta distribution}
\usage{
mcmap_beta(
  target,
  method = "",
  integrate_controls = list(),
  optim_controls = list()
)
}
\arguments{
\item{target}{A vector of size 2. The first element is mean and the second element is c-statistic.}

\item{method}{Not implemented for this funciton yet; leave as empty string.}

\item{integrate_controls}{(optional): parameters to be passed to integrate()}

\item{optim_controls}{(optional): parameters to be passed to optim()}
}
\value{
A vector of size two that contains the distribution parameters
}
\description{
Maps a pair of mean and c-statistic value to the parameters of a beta distribution
}
\examples{
mcmap_beta(c(0.1, 0.75))
}
