% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mcmap_generic}
\alias{mcmap_generic}
\title{A generic mapper function}
\usage{
mcmap_generic(
  target,
  CDF,
  integrate_controls = list(),
  optim_controls = list()
)
}
\arguments{
\item{target}{A vector of size 2. The first element is mean and the second element is c-statistic.}

\item{CDF}{Cumulative distribution function of an unnspecified distribution. The CDF must be indexed by two parameters.}

\item{integrate_controls}{(optional): parameters to be passed to integrate()}

\item{optim_controls}{(optional): parameters to be passed to optim()}
}
\value{
A vector of size two that contains the distribution parameters
}
\description{
Maps a pair of mean and c-statistic value to the parameters of an unspecified distribution that is indexed by two parameters
}
\examples{
mcmap_generic(c(0.1, 0.75), pbeta)
}
