% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitnorm.R
\name{probitnorm}
\alias{probitnorm}
\alias{dprobitnorm}
\alias{pprobitnorm}
\alias{rprobitnorm}
\alias{qprobitnorm}
\title{Functions related to probit-normal distribution.}
\usage{
dprobitnorm(x, mu, sigma)

pprobitnorm(x, mu, sigma)

rprobitnorm(n, mu, sigma)

qprobitnorm(x, mu, sigma)
}
\arguments{
\item{x}{For density, CDF, and quantile functions}

\item{mu}{Mean of the probit-transformed variable}

\item{sigma}{SD of the probit-transformed variable}

\item{n}{Number of draws requested (for rprobitnorm)}
}
\value{
Depends on the function
}
\description{
Functions related to probit-normal distribution.
}
